+++ packages/core/src/index.ts
@@ -1,127 +1,125 @@
-import { useRowVirtualizer } from "./table/hooks/use_row_virtualizer";
-import { ReactTanstackTableUi } from "./table/table";
-import { iterateOverColumns } from "./iterate_over_columns";
-import type { Skin } from "./skin";
-import { useTableContext } from "./table/table_context";
-import { useColVirtualizer } from "./table/hooks/use_col_virtualizer";
-import type { VirtualHeaderCell } from "./table/types";
-import { useTableCssVars } from "./skin";
-import {
-  defaultSkin,
-  darkModeVars,
-  lightModeVars,
-} from "./default_skin/default_skin";
-import { HeaderPinButtons } from "./default_skin/header_pin_buttons";
-import { Checkbox } from "./default_skin/checkbox";
+import { Cell, type CellProps } from "./default_skin/cell";
 import { CellAvatar } from "./default_skin/cell_avatar";
 import { CellAvatarWithText } from "./default_skin/cell_avatar_with_text";
-import { CellNumber } from "./default_skin/cell_number";
-import { CellTag } from "./default_skin/cell_tag";
-import { CellText } from "./default_skin/cell_text";
-import { Resizer } from "./default_skin/resizer";
-import { ExpandButton } from "./default_skin/expand_button";
-import { RowPinButtons } from "./default_skin/row_pin_buttons";
-import { CellTextBold } from "./default_skin/cell_text_bold";
-import { Header } from "./default_skin/header";
+import {
+  CellBadge,
+  type BadgeColor,
+  type CellBadgeProps,
+} from "./default_skin/cell_badge";
 import {
   CellCurrency,
   type CellCurrencyProps,
 } from "./default_skin/cell_currency";
-import {
-  CellBadge,
-  type CellBadgeProps,
-  type BadgeColor,
-} from "./default_skin/cell_badge";
-import { Cell, type CellProps } from "./default_skin/cell";
 import { CellLink, type CellLinkProps } from "./default_skin/cell_link";
+import { CellNumber } from "./default_skin/cell_number";
 import {
   CellPercent,
   type CellPercentProps,
 } from "./default_skin/cell_percent";
+import { CellTag } from "./default_skin/cell_tag";
+import { CellText } from "./default_skin/cell_text";
+import { CellTextBold } from "./default_skin/cell_text_bold";
+import { Checkbox } from "./default_skin/checkbox";
+import {
+  darkModeVars,
+  defaultSkin,
+  lightModeVars,
+} from "./default_skin/default_skin";
+import { ExpandButton } from "./default_skin/expand_button";
+import { Header } from "./default_skin/header";
+import { HeaderPinButtons } from "./default_skin/header_pin_buttons";
+import { Resizer } from "./default_skin/resizer";
+import { RowPinButtons } from "./default_skin/row_pin_buttons";
+import { iterateOverColumns } from "./iterate_over_columns";
+import type { Skin } from "./skin";
+import { useTableCssVars } from "./skin";
+import { ReactTanstackTableUi } from "./table/table";
+import { useTableContext } from "./table/table_context";
+import type { VirtualHeaderCell } from "./table/types";
 
+import type { AutoSizerProps } from "./auto_sizer";
+import { AutoSizer } from "./auto_sizer";
+import type { CellDecorator, HeaderDecorator } from "./decorate_column_helper";
+import { decorateColumnHelper } from "./decorate_column_helper";
+import type { CellAvatarProps } from "./default_skin/cell_avatar";
+import type { CellAvatarWithTextProps } from "./default_skin/cell_avatar_with_text";
+import type { CheckboxProps } from "./default_skin/checkbox";
+import { useAutoSizer } from "./table/hooks/use_autosizer";
+import { useCellProps } from "./table/hooks/use_cell_props";
 import { useColProps } from "./table/hooks/use_col_props";
 import { useColRef } from "./table/hooks/use_col_ref";
-import { useCellProps } from "./table/hooks/use_cell_props";
-import { useRowRef } from "./table/hooks/use_row_ref";
-import { useRowProps } from "./table/hooks/use_row_props";
-import { useCrushHeader } from "./table/hooks/use_crush_header";
 import { useCrushAllCols } from "./table/hooks/use_crush_all_cols";
-import { decorateColumnHelper } from "./decorate_column_helper";
-import type { HeaderDecorator, CellDecorator } from "./decorate_column_helper";
-import type { RttuiRef } from "./table/types";
-import { createTablePropsSelector, shallowEqual, strictEqual } from "./utils";
+import { useCrushHeader } from "./table/hooks/use_crush_header";
+import { useRowProps } from "./table/hooks/use_row_props";
+import { useRowRef } from "./table/hooks/use_row_ref";
 import {
   useListenToTableProps,
   useTableProps,
 } from "./table/hooks/use_table_props";
-import type { CheckboxProps } from "./default_skin/checkbox";
-import type { CellAvatarProps } from "./default_skin/cell_avatar";
-import type { CellAvatarWithTextProps } from "./default_skin/cell_avatar_with_text";
-import type { ReactTanstackTableUiProps } from "./table/types";
-import { AutoSizer } from "./auto_sizer";
-import type { AutoSizerProps } from "./auto_sizer";
-import type { CrushBy } from "./table/types";
-import { useAutoSizer } from "./table/hooks/use_autosizer";
+import type {
+  CrushBy,
+  ReactTanstackTableUiProps,
+  RttuiRef,
+} from "./table/types";
+import { createTablePropsSelector, shallowEqual, strictEqual } from "./utils";
 
 export {
-  useListenToTableProps,
-  ReactTanstackTableUi,
-  useRowVirtualizer,
-  iterateOverColumns,
-  useRowProps,
-  useTableContext,
-  useColVirtualizer,
-  useTableCssVars,
-  defaultSkin,
-  darkModeVars,
-  lightModeVars,
-  useColProps,
-  createTablePropsSelector,
-  useCrushHeader,
-  decorateColumnHelper,
-  useTableProps,
-  useColRef,
-  useRowRef,
-  useCellProps,
-  shallowEqual,
-  strictEqual,
-  HeaderPinButtons,
-  Checkbox,
+  AutoSizer,
+  Cell,
   CellAvatar,
   CellAvatarWithText,
+  CellBadge,
+  CellCurrency,
+  CellLink,
   CellNumber,
+  CellPercent,
   CellTag,
   CellText,
-  Resizer,
-  ExpandButton,
-  RowPinButtons,
   CellTextBold,
-  CellCurrency,
-  CellBadge,
-  Cell,
-  CellLink,
-  CellPercent,
-  AutoSizer,
+  Checkbox,
+  createTablePropsSelector,
+  darkModeVars,
+  decorateColumnHelper,
+  defaultSkin,
+  ExpandButton,
   Header,
-  useCrushAllCols,
+  HeaderPinButtons,
+  iterateOverColumns,
+  lightModeVars,
+  ReactTanstackTableUi,
+  Resizer,
+  RowPinButtons,
+  shallowEqual,
+  strictEqual,
   useAutoSizer,
+  useCellProps,
+  useColProps,
+  useColRef,
+  useCrushAllCols,
+  useCrushHeader,
+  useListenToTableProps,
+  useRowProps,
+  useRowRef,
+  useTableContext,
+  useTableCssVars,
+  useTableProps,
 };
 export type {
-  CrushBy,
   AutoSizerProps,
-  Skin,
-  VirtualHeaderCell,
-  HeaderDecorator,
-  CellDecorator,
-  RttuiRef,
-  CheckboxProps,
+  BadgeColor,
   CellAvatarProps,
   CellAvatarWithTextProps,
-  ReactTanstackTableUiProps,
   CellBadgeProps,
-  BadgeColor,
   CellCurrencyProps,
-  CellProps,
+  CellDecorator,
   CellLinkProps,
   CellPercentProps,
+  CellProps,
+  CheckboxProps,
+  CrushBy,
+  HeaderDecorator,
+  ReactTanstackTableUiProps,
+  RttuiRef,
+  Skin,
+  VirtualHeaderCell,
 };