+++ packages/core/src/table/contexts/col_virtualizer_context.tsx
@@ -1,7 +1,0 @@
-"use client";
-import React from "react";
-import { ColVirtualizerType } from "../cols/col_virtualizer_type";
-
-export const ColVirtualizerContext = React.createContext<
-  ColVirtualizerType | undefined
->(undefined);