+++ packages/core/src/table/contexts/row_virtualizer_context.tsx
@@ -1,7 +1,0 @@
-"use client";
-import React from "react";
-import { RowVirtualizerContextType } from "../rows/row_virtualizer_context_type";
-
-export const RowVirtualizerContext = React.createContext<
-  RowVirtualizerContextType | undefined
->(undefined);