+++ nextjs-test/CHANGELOG.md
@@ -1,5 +1,12 @@
 # nextjs-test
 
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @rttui/core@1.0.52
+
 ## 0.1.7
 
 ### Patch Changes